Create and rename files for logic lab
